InitalCommit
Added Maze Game and Ware Game c++ files as an example of work. Wargame is debugged as best I could could not find any more bugs in it and checked for memory leaks. 
MazeGame is a work in progress that has hit the backburner. works and is operational but has more potential and learning hidden inside by adding more functionality